Working on Workouts
Created a reccomended routine workout that can parse a text file and
create a workout with warmups and sets of progressions with exercises.
Also added a .gitignore for testing. Need to parse from a pebble
resource file next. Also, should there be a progression routine
for a group of progressions (skill and strength)?